Fix: decouple AnalysisResultsCard
